Add partition to death stats before and while covid, blik data and flu data.

+++ Covid_App/Entities/BlikPayments.cs
@@ -5,7 +5,7 @@
         public string Kwartal { get; set; }
         public string Wartosc { get; set; }
         public string Proc_BLIK { get; set; }
-        public long Liczba { get; set; }
+        public int Liczba { get; set; }
         public string Sr_dzienna_liczba_platnosci { get; set; }
         public string Sr_wartosc_płatnosci { get; set; }
     }
