+++ Covid_App/Entities/FluData.cs
@@ -3,7 +3,7 @@
     public class FluData
     {
         public int Rok { get; set; }
-        public long Liczba { get; set; }
+        public int Liczba { get; set; }
         public double Zapadalnosc { get; set; }
         public double Spadek_wzrost_zapadalnosci { get; set; }
     }
